Add course completion rules and score reporting
There is now course completion rules for the entire course instead of
just for page. You can set the total time for the course to meet certain
accrediation. You can also set a minimum grade and allow students to get
a failure in the grade book. If the course is more of a 'checked' you
did it kind of deal, all you have to do is make completeOnly true in the
JSON file.