Add encoding
Saved data has been shorten to only changed values (delta) and
compressed into a single array of mostly numbers. Given the same schema,
the serialize and unserialize functions are able to take the delta
object, flatten it, compress it with a help from the telemetry object,
call state.save, and then do it all in reverse. Inital tests show that
when the user reaches the example video page, the orginal save file was
3887 chars. The new save file (with encoding and gzip compression) is
     180 chars for the same progress. Computers are amazing.