Add ability to send analytics
After testing the SCORM package, it seems that a lot of LMS programs do
not show saved or reported data besides time on lesson and grade. To get
around this, the ability to send the data to a Google Form has been
added. You need the Google Form ID, and two text boxes with their IDs.